provide default symbol processors
