+++ build.config.ts
@@ -1,5 +1,6 @@
 import { defineBuildConfig } from 'unbuild'
 
 export default defineBuildConfig({
+  entries: ['./src/processors.ts'],
   externals: ['consola', 'ohash', 'pathe', 'ufo', 'h3'],
 })
